requirements.txt
fastapi
httpx
uvicorn
python-multipart
pillow
